+++ tests/test_eventloop.py
@@ -10,7 +10,7 @@ from pytest_mock.plugin import MockerFixture
 from anyio import run, sleep_forever, sleep_until
 
 if sys.version_info < (3, 8):
-    from mock import AsyncMock
+    from unittest.mock import AsyncMock
 else:
     from unittest.mock import AsyncMock